moved mentions detection function into new mentions service, moved tests also
